Reproduce NoSuchElementException when storing empty serialized items twice in a row
Added a new storage class `MyStorage` to manage a list of `MyItemDto` objects.
This class utilizes Multiplatform Settings for persistence and kotlinx.serialization for data serialization.
The functionality includes getting and replacing the list of items.

The `MainActivity` has been updated with functions to demonstrate the usage of the storage, including storing and retrieving empty and non-empty lists of items.

The necessary dependencies for Multiplatform Settings and kotlinx.serialization have been added to the project.